fix: improve media display in chat media screen
